refactor/structure made
